+++ utils/utils.go
@@ -23,6 +23,8 @@ import (
 	"github.com/cossacklabs/themis/gothemis/keys"
 	"io"
 	"io/ioutil"
+	"sync"
+	"time"
 
 	log "github.com/sirupsen/logrus"
 	"os"
@@ -194,4 +196,20 @@ func Min(x, y int) int {
 // GetConfigPathByName returns filepath to config file named "name" from default configs folder
 func GetConfigPathByName(name string) string {
 	return fmt.Sprintf("configs/%s.yaml", name)
+}
+
+// WaitWithTimeout waits for the waitgroup for the specified max timeout.
+// Returns true if waiting timed out.
+func WaitWithTimeout(wg *sync.WaitGroup, timeout time.Duration) bool {
+	c := make(chan struct{})
+	go func() {
+		defer close(c)
+		wg.Wait()
+	}()
+	select {
+	case <-c:
+		return false // completed normally
+	case <-time.After(timeout):
+		return true // timed out
+	}
 }
